:recycle: Flyttet outside-click handling ut av hver hook
